+++ src/ts/dom.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        dom.ts
- * @version     v1.2.0
+ * @version     v1.3.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -26,7 +26,7 @@ export namespace DomElement {
 
         result.setAttribute( Constant.PEEK_JS_IGNORE_STATE_ATTRIBUTE, IgnoreState.ignore );
 
-        if ( Is.defined( className ) ) {
+        if ( Is.definedString( className ) ) {
             result.className = className;
         }
 
@@ -45,7 +45,7 @@ export namespace DomElement {
 
     export function cancelBubble( e: Event ) {
         e.preventDefault();
-        e.cancelBubble = true;
+        e.stopPropagation();
     }
 
     export function getScrollPosition() : Position {
@@ -59,7 +59,7 @@ export namespace DomElement {
         return result;
     }
 
-    export function showElementAtMousePosition( e: any, element: HTMLElement ) {
+    export function showElementAtMousePosition( e: MouseEvent, element: HTMLElement ) {
         if ( element.style.display !== "block" ) {
             let left: number = e.pageX;
             let top: number = e.pageY;