+++ src/ts/enum.ts
@@ -4,7 +4,7 @@
  * A lightweight JavaScript library that attaches a viewer to a specific node type, allowing you to view the CSS properties, attributes, and size/position.
  * 
  * @file        enum.ts
- * @version     v1.2.0
+ * @version     v1.3.0
  * @author      Bunoon
  * @license     MIT License
  * @copyright   Bunoon 2024
@@ -21,6 +21,7 @@ export const enum Mode {
     css = 1,
     attributes = 2,
     size = 3,
+    class = 4,
 }
 
 export const enum Value {